requirements.txt
flask
llama-index
streamlit
Ollama
nltk
indic-nlp-library
transformers
torch